requirements.txt
plyfile
ipython
numpy==1.26.4
matplotlib
Pillow
imageio
imageio-ffmpeg
rembg[gpu]
lpips
einops
safetensors
sentencepiece
accelerate
transformers
diffusers==0.32
kiui
omegaconf
pyarrow
redis
loguru
pandas
image-reward
git+https://github.com/openai/CLIP.git
deepspeed==0.15.4
icecream
trimesh
open3d
shapely
jaxtyping==0.2.23
tensorboard==2.12.0
wandb
torchmetrics
mapbox-earcut
recordclass
diptest==0.6.1
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.1+cu121
torchvision==0.18.1+cu121
xformers==0.0.27
iopath
pytorch3d